requirements.txt
accelerate
peft
transformers
trl
datasets
PyYAML
pydantic_settings
bitsandbytes